// src/signs/generated/artwork/M1-5.ts
// Generated from official FHWA 2024 SHS artwork. Do not edit.
import type { SignArt } from "../../artwork";

export const ART_M1_5: SignArt = {
  "width": 24,
  "height": 24,
  "paths": [
    {
      "d": "M1.5 0L22.5 0C23.328 0 24 0.672 24 1.5L24 22.5C24 23.328 23.328 24 22.5 24L1.5 24C0.672 24 0 23.328 0 22.5L0 1.5C0 0.672 0.672 0 1.5 0Z",
      "fill": "black"
    },
    {
      "d": "M23.5 12C23.5 18.351 18.351 23.5 12 23.5C5.649 23.5 0.5 18.351 0.5 12C0.5 5.649 5.649 0.5 12 0.5C18.351 0.5 23.5 5.649 23.5 12Z",
      "fill": "white"
    }
  ],
  "legend": [
    {
      "d": "M16.479 18L16.479 8.274L15.396 8.274L15.396 6.758L15.521 6.758C16.021 6.758 16.562 6.463 16.896 6L18.396 6L18.396 18L16.479 18Z",
      "fill": "black"
    },
    {
      "d": "M11.1 12.382C9.485 13.387 7.998 13.831 7.1 16.063L12.996 16.063L12.996 18L4.827 18C4.869 16.611 5.159 15.471 5.837 14.337C7.943 10.817 11.101 11.206 11.101 9.326C11.101 8.442 10.09 7.726 8.911 7.726C7.69 7.726 6.932 8.316 6.721 9.411L4.827 8.99C5.248 7.011 6.806 5.789 8.911 5.789C11.185 5.789 12.996 7.389 12.996 9.368C12.996 10.803 12.17 11.717 11.1 12.382Z",
      "fill": "black"
    }
  ]
};